requirements.txt
requests==2.25.1
urllib3==1.25.9
Pillow==8.3.2
python-dotenv==0.19.0
python_dateutil==2.8.2
